+++ packages/iov-bns/src/testdata.spec.ts
@@ -129,6 +129,7 @@ const randomMsg: SendTransaction & WithCreator = {
       fractionalDigits: 9,
       tokenTicker: "PSQL" as TokenTicker,
     },
+    payer: address,
   },
 };